+++ app/screens/people-screen/contacts/contact-transactions.tsx
@@ -2,8 +2,8 @@ import { gql } from "@apollo/client"
 import { useI18nContext } from "@app/i18n/i18n-react"
 import * as React from "react"
 import { SectionList, Text, View } from "react-native"
-import { TransactionItem } from "../../components/transaction-item"
-import { toastShow } from "../../utils/toast"
+import { TransactionItem } from "../../../components/transaction-item"
+import { toastShow } from "../../../utils/toast"
 
 import { useTransactionListForContactQuery } from "@app/graphql/generated"
 import { useIsAuthed } from "@app/graphql/is-authed-context"
