+++ app/screens/people-screen/contacts/contacts-card.tsx
