+++ app/screens/people-screen/people-screen.types.d.ts
